# feat(#35): implement ColumnWriter interface for direct binary → Arrow conversion