--- a/types.go
+++ b/types.go
@@ -6,6 +6,7 @@ import (
 	"math"
 
 	"github.com/apache/arrow-go/v18/arrow"
+	"github.com/apache/arrow-go/v18/arrow/array"
 )
 
 const (
@@ -26,6 +27,22 @@ type TypeHandler interface {
 	// Parse converts binary PostgreSQL data to Go value
 	// Returns nil for NULL values (empty data)
 	Parse(data []byte) (any, error)
+}
+
+// ColumnWriter defines the interface for direct binary data → Arrow column conversion
+// without intermediate allocations. This replaces the TypeHandler.Parse pattern
+// for performance-critical scenarios.
+type ColumnWriter interface {
+	// WriteField writes binary PostgreSQL data directly to Arrow column
+	// data: binary PostgreSQL field data
+	// isNull: true if field is NULL (data should be ignored)
+	WriteField(data []byte, isNull bool) error
+
+	// ArrowType returns the corresponding Arrow data type
+	ArrowType() arrow.DataType
+
+	// BuilderStats returns current length and capacity for optimization insights
+	BuilderStats() (length, capacity int)
 }
 
 // TypeRegistry manages type handlers by OID
@@ -555,4 +572,359 @@ func (t *IntervalType) Parse(data []byte) (any, error) {
 		Days:        pgDays,
 		Nanoseconds: arrowNanos,
 	}, nil
+}
+
+// ColumnWriter implementations for direct binary → Arrow conversion
+
+// BoolColumnWriter writes PostgreSQL bool data directly to Arrow boolean arrays
+type BoolColumnWriter struct {
+	Builder *array.BooleanBuilder
+}
+
+func (w *BoolColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 1 {
+		return fmt.Errorf("invalid data length for bool: expected 1, got %d", len(data))
+	}
+
+	w.Builder.Append(data[0] != 0)
+	return nil
+}
+
+func (w *BoolColumnWriter) ArrowType() arrow.DataType {
+	return arrow.FixedWidthTypes.Boolean
+}
+
+func (w *BoolColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// Int16ColumnWriter writes PostgreSQL int2 data directly to Arrow int16 arrays
+type Int16ColumnWriter struct {
+	Builder *array.Int16Builder
+}
+
+func (w *Int16ColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 2 {
+		return fmt.Errorf("invalid data length for int16: expected 2, got %d", len(data))
+	}
+
+	value := int16(binary.BigEndian.Uint16(data))
+	w.Builder.Append(value)
+	return nil
+}
+
+func (w *Int16ColumnWriter) ArrowType() arrow.DataType {
+	return arrow.PrimitiveTypes.Int16
+}
+
+func (w *Int16ColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// Int32ColumnWriter writes PostgreSQL int4 data directly to Arrow int32 arrays
+type Int32ColumnWriter struct {
+	Builder *array.Int32Builder
+}
+
+func (w *Int32ColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 4 {
+		return fmt.Errorf("invalid data length for int32: expected 4, got %d", len(data))
+	}
+
+	value := int32(binary.BigEndian.Uint32(data))
+	w.Builder.Append(value)
+	return nil
+}
+
+func (w *Int32ColumnWriter) ArrowType() arrow.DataType {
+	return arrow.PrimitiveTypes.Int32
+}
+
+func (w *Int32ColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// Int64ColumnWriter writes PostgreSQL int8 data directly to Arrow int64 arrays
+type Int64ColumnWriter struct {
+	Builder *array.Int64Builder
+}
+
+func (w *Int64ColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 8 {
+		return fmt.Errorf("invalid data length for int64: expected 8, got %d", len(data))
+	}
+
+	value := int64(binary.BigEndian.Uint64(data))
+	w.Builder.Append(value)
+	return nil
+}
+
+func (w *Int64ColumnWriter) ArrowType() arrow.DataType {
+	return arrow.PrimitiveTypes.Int64
+}
+
+func (w *Int64ColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// Float32ColumnWriter writes PostgreSQL float4 data directly to Arrow float32 arrays
+type Float32ColumnWriter struct {
+	Builder *array.Float32Builder
+}
+
+func (w *Float32ColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 4 {
+		return fmt.Errorf("invalid data length for float32: expected 4, got %d", len(data))
+	}
+
+	value := math.Float32frombits(binary.BigEndian.Uint32(data))
+	w.Builder.Append(value)
+	return nil
+}
+
+func (w *Float32ColumnWriter) ArrowType() arrow.DataType {
+	return arrow.PrimitiveTypes.Float32
+}
+
+func (w *Float32ColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// Float64ColumnWriter writes PostgreSQL float8 data directly to Arrow float64 arrays
+type Float64ColumnWriter struct {
+	Builder *array.Float64Builder
+}
+
+func (w *Float64ColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 8 {
+		return fmt.Errorf("invalid data length for float64: expected 8, got %d", len(data))
+	}
+
+	value := math.Float64frombits(binary.BigEndian.Uint64(data))
+	w.Builder.Append(value)
+	return nil
+}
+
+func (w *Float64ColumnWriter) ArrowType() arrow.DataType {
+	return arrow.PrimitiveTypes.Float64
+}
+
+func (w *Float64ColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// StringColumnWriter writes PostgreSQL text data directly to Arrow string arrays
+// Achieves zero-copy by appending bytes directly to the binary builder
+type StringColumnWriter struct {
+	Builder *array.StringBuilder
+}
+
+func (w *StringColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	// Zero-copy: append bytes directly without string conversion
+	w.Builder.BinaryBuilder.Append(data)
+	return nil
+}
+
+func (w *StringColumnWriter) ArrowType() arrow.DataType {
+	return arrow.BinaryTypes.String
+}
+
+func (w *StringColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// Date32ColumnWriter writes PostgreSQL date data directly to Arrow date32 arrays
+type Date32ColumnWriter struct {
+	Builder *array.Date32Builder
+}
+
+func (w *Date32ColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 4 {
+		return fmt.Errorf("invalid data length for date: expected 4, got %d", len(data))
+	}
+
+	pgDays := int32(binary.BigEndian.Uint32(data))
+	arrowDays := pgDays + PostgresDateEpochDays // Single addition operation
+	w.Builder.Append(arrow.Date32(arrowDays))
+	return nil
+}
+
+func (w *Date32ColumnWriter) ArrowType() arrow.DataType {
+	return arrow.PrimitiveTypes.Date32
+}
+
+func (w *Date32ColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// Time64ColumnWriter writes PostgreSQL time data directly to Arrow time64 arrays
+type Time64ColumnWriter struct {
+	Builder *array.Time64Builder
+}
+
+func (w *Time64ColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 8 {
+		return fmt.Errorf("invalid data length for time: expected 8, got %d", len(data))
+	}
+
+	// PostgreSQL time is stored as microseconds since midnight
+	// Arrow Time64[microsecond] also uses microseconds - direct conversion
+	timeMicros := int64(binary.BigEndian.Uint64(data))
+	w.Builder.Append(arrow.Time64(timeMicros))
+	return nil
+}
+
+func (w *Time64ColumnWriter) ArrowType() arrow.DataType {
+	return arrow.FixedWidthTypes.Time64us
+}
+
+func (w *Time64ColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// TimestampColumnWriter writes PostgreSQL timestamp data directly to Arrow timestamp arrays
+type TimestampColumnWriter struct {
+	Builder       *array.TimestampBuilder
+	timestampType *arrow.TimestampType
+}
+
+func (w *TimestampColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 8 {
+		return fmt.Errorf("invalid data length for timestamp: expected 8, got %d", len(data))
+	}
+
+	// PostgreSQL timestamp is stored as microseconds since 2000-01-01
+	// Arrow timestamp uses microseconds since 1970-01-01
+	// Add epoch adjustment to convert from PostgreSQL epoch to Arrow epoch
+	// IMPORTANT: Use signed int64 conversion, not uint64, since PG timestamps can be negative (before 2000-01-01)
+	pgMicros := int64(binary.BigEndian.Uint64(data))
+	arrowMicros := pgMicros + PostgresTimestampEpochMicros
+	w.Builder.Append(arrow.Timestamp(arrowMicros))
+	return nil
+}
+
+func (w *TimestampColumnWriter) ArrowType() arrow.DataType {
+	return w.timestampType
+}
+
+func (w *TimestampColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
+}
+
+// NewTimestampColumnWriter creates a TimestampColumnWriter for timestamp (no timezone)
+func NewTimestampColumnWriter(builder *array.TimestampBuilder) *TimestampColumnWriter {
+	return &TimestampColumnWriter{
+		Builder:       builder,
+		timestampType: &arrow.TimestampType{Unit: arrow.Microsecond, TimeZone: ""},
+	}
+}
+
+// NewTimestamptzColumnWriter creates a TimestampColumnWriter for timestamptz (UTC timezone)
+func NewTimestamptzColumnWriter(builder *array.TimestampBuilder) *TimestampColumnWriter {
+	return &TimestampColumnWriter{
+		Builder:       builder,
+		timestampType: &arrow.TimestampType{Unit: arrow.Microsecond, TimeZone: "UTC"},
+	}
+}
+
+// MonthDayNanoIntervalColumnWriter writes PostgreSQL interval data directly to Arrow interval arrays
+type MonthDayNanoIntervalColumnWriter struct {
+	Builder *array.MonthDayNanoIntervalBuilder
+}
+
+func (w *MonthDayNanoIntervalColumnWriter) WriteField(data []byte, isNull bool) error {
+	if isNull {
+		w.Builder.AppendNull()
+		return nil
+	}
+
+	if len(data) != 16 {
+		return fmt.Errorf("invalid data length for interval: expected 16, got %d", len(data))
+	}
+
+	// PostgreSQL interval binary format (16 bytes):
+	// Bytes 0-7:   int64 microseconds (time field)
+	// Bytes 8-11:  int32 days
+	// Bytes 12-15: int32 months
+	// IMPORTANT: Use signed int64 conversion, not uint64, since intervals can be negative
+	pgMicros := int64(binary.BigEndian.Uint64(data[0:8]))
+	pgDays := int32(binary.BigEndian.Uint32(data[8:12]))
+	pgMonths := int32(binary.BigEndian.Uint32(data[12:16]))
+
+	// Convert microseconds to nanoseconds with overflow check
+	const maxMicros = math.MaxInt64 / 1000
+	const minMicros = math.MinInt64 / 1000
+	if pgMicros > maxMicros || pgMicros < minMicros {
+		return fmt.Errorf("interval microseconds overflow: %d", pgMicros)
+	}
+	arrowNanos := pgMicros * 1000
+
+	// Return Arrow MonthDayNanoInterval with field reordering:
+	// PostgreSQL: (microseconds, days, months)
+	// Arrow: (months, days, nanoseconds)
+	interval := arrow.MonthDayNanoInterval{
+		Months:      pgMonths,
+		Days:        pgDays,
+		Nanoseconds: arrowNanos,
+	}
+	w.Builder.Append(interval)
+	return nil
+}
+
+func (w *MonthDayNanoIntervalColumnWriter) ArrowType() arrow.DataType {
+	return arrow.FixedWidthTypes.MonthDayNanoInterval
+}
+
+func (w *MonthDayNanoIntervalColumnWriter) BuilderStats() (length, capacity int) {
+	return w.Builder.Len(), w.Builder.Cap()
 }